controllers auth, refactoring of packages
