+++ doc-control-app-spring/src/main/java/it/akademija/wizards/controllers/UserGroupController.java
@@ -7,11 +7,13 @@ import it.akademija.wizards.models.usergroup.*;
 import it.akademija.wizards.services.UserGroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @Api(value = "user groups")
+@PreAuthorize("hasRole('ADMIN')")
 @RestController
 @RequestMapping(value = "/api/groups")
 public class UserGroupController {
