+++ doc-control-app-spring/src/main/java/it/akademija/wizards/security/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package it.akademija.wizards.security;
+package it.akademija.wizards.security.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
