+++ doc-control-app-spring/src/main/java/it/akademija/wizards/security/jwt/JwtTokenProvider.java
@@ -1,6 +1,7 @@
-package it.akademija.wizards.security;
+package it.akademija.wizards.security.jwt;
 
 import io.jsonwebtoken.*;
+import it.akademija.wizards.security.models.UserPrincipal;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
