+++ doc-control-app-spring/src/main/java/it/akademija/wizards/security/payload/ApiResponse.java
@@ -1,4 +1,4 @@
-package it.akademija.wizards.payload;
+package it.akademija.wizards.security.payload;
 
 public class ApiResponse {
     private Boolean success;
